Fix issue with watching mounted folders from within a docker container
This commit hacks a fix for a bizarre problem with obtaining fsnotify
events from within nested folders that are within an osxfs mount running
in Docker.

It turns out that fsnotify was not picking up any events from inotify
when a deeply nested folder was being watched but none of that folder's
parents. The hacky fix for this is to detect, while running in docker,
if we're about to watch a folder within an fuse.osxfs mount and
recursively watch all parent folders.